+++ discovery.go
@@ -75,7 +75,7 @@ func (r *Router) BucketDiscovery(ctx context.Context, bucketID uint64) (*Replica
 		rsID := rsID
 		go func(_rs *Replicaset) {
 			defer wg.Done()
-			_, errStat := _rs.bucketStat(ctx, bucketID)
+			_, errStat := _rs.BucketStat(ctx, bucketID)
 			if errStat == nil {
 				resultRs, err = r.BucketSet(bucketID, rsID)
 			}